Isolate the CPU part and use OpenCV to read and access images
