+++ draw_edges_by_matlab/draw_edges_from_list.m
@@ -1,22 +1,20 @@
 clc; close all;
 
-dst_path = '/home/chchien/BrownU/research/Differential-Geometry-Edge-Detection/MyBitBucket/canny/';
-output_data_path = 'test_files/';
+dst_path = '/gpfs/data/bkimia/cchien3/Third-Order-Edge-Detector/';
+output_data_path = 'output_files/';
 
 % -- read image to retrieve image height and width --
 input_img_folder = 'input_images/';
-input_img_name = '2018';
-str_readPath = strcat(dst_path, input_img_folder, input_img_name, '.pgm');
+input_img_name = 'euroc_sample_img';
+str_readPath = strcat(dst_path, input_img_folder, input_img_name, '.png');
 img = imread(str_readPath);
 img_width = size(img,2);
 img_height = size(img,1);
 
 % -- read edge map text files --
-edge_list_pts_file = 'data_final_output.txt';
+edge_list_pts_file = 'data_final_output_cpu.txt';
 full_edge_file = fullfile(dst_path, output_data_path, edge_list_pts_file);
-edge_map = fopen(full_edge_file, 'r');
-ldata = textscan(edge_map, '%f\t%f\t%f\t%f', 'CollectOutput', true );
-pixel2d = ldata{1,1};
+toed_edges = importdata(full_edge_file);
 
 % % -- strong edges --
 % edge_strong_pts_file = 'data_strong.txt';
@@ -33,10 +31,8 @@ pixel2d = ldata{1,1};
 % pixel2d_weak = ldata_weak{1,1};
 
 figure;
-plot(pixel2d(:,1),pixel2d(:,2),'.', 'Color', [0.1249  0.7851  0.6067]);
-xlim([-50, 350]);
-ylim([-50, 500]);
-axis equal;
+imshow(img); hold on;
+scatter(toed_edges(:,1), toed_edges(:,2), 'c.');
 set(gcf,'color','w');
 
 % figure;
